Make console output persistent by moving it out to DevelopmentMode.tsx, compilation errors are now displayed in console
