+++ client/src/components/DevelopmentMode/Console.tsx
@@ -3,30 +3,9 @@ import { socket } from 'utils/socket';
 
 import styles from 'styles/Console.module.css';
 
-const Console = () => {
-  const [chunks, setChunks] = useState<string[]>([]);
+const Console = ({ chunks, handleAddChunk, scrollToBottom }) => {
   const [input, setInput] = useState('');
   const inputElement = useRef(null);
-
-  const scrollToBottom = () => {
-    if (inputElement.current) {
-      const container = inputElement.current.parentElement;
-      container.scrollTop = container.scrollHeight;
-    }
-  };
-
-  useEffect(() => {
-    const onStdout = (data: string) => {
-      setChunks((oldChunks) => [...oldChunks, data]);
-      scrollToBottom();
-    };
-
-    socket.on('sendStdoutToUser', onStdout);
-
-    return () => {
-      socket.off('sendStdoutToUser', onStdout);
-    };
-  }, []);
 
   const handleInput = () => {
     setInput(inputElement.current.innerText);
@@ -41,7 +20,7 @@ const Console = () => {
     if (event.key === 'Enter') {
       if (input.length > 0) {
         socket.emit('send_stdin', input);
-        setChunks([...chunks, `${input}\n`]);
+        handleAddChunk(`${input}\n`);
         clearInput();
         scrollToBottom();
       }
